+++ SqlSugar.Attributes.Test/Entities/Role.cs
@@ -1,20 +1,23 @@
 ﻿namespace SqlSugar.Attributes.Test.Entities
 {
+    /// <summary>
+    /// Role
+    /// </summary>
     [SugarTable("Role")]
     public class Role
     {
         /// <summary>
-        /// 角色Id
+        /// Role Id
         /// </summary>
         [SugarColumn(IsIdentity = true, IsPrimaryKey = true)]
-        public long RoleId { get; set; }
+        public int RoleId { get; set; }
         /// <summary>
-        /// 角色名
+        /// Role name
         /// </summary>
         [SugarColumn(Length = 50, IsNullable = false)]
         public string RoleName { get; set; }
         /// <summary>
-        /// 创建日期
+        /// Create time
         /// </summary>
         [SugarColumn(IsNullable = false, InsertServerTime = true)]
         public DateTime CreateTime { get; set; }
