+++ SqlSugar.Attributes.Test/Entities/User.cs
@@ -1,13 +1,16 @@
 ﻿namespace SqlSugar.Attributes.Test.Entities
 {
+    /// <summary>
+    /// User
+    /// </summary>
     [SugarTable("User")]
     public class User
     {
         /// <summary>
-        /// 用户Id
+        /// User Id
         /// </summary>
         [SugarColumn(IsIdentity = true, IsPrimaryKey = true)]
-        public long UserId { get; set; }
+        public int UserId { get; set; }
         /// <summary>
         /// User name
         /// 用户名
@@ -15,12 +18,12 @@
         [SugarColumn(Length = 50, IsNullable = false)]
         public string UserName { get; set; }
         /// <summary>
-        /// 用户邮箱
+        /// Email
         /// </summary>
         [SugarColumn(IsNullable = true)]
         public string Email { get; set; }
         /// <summary>
-        /// 创建日期
+        /// Create time
         /// </summary>
         [SugarColumn(IsNullable = false)]
         public DateTime CreateTime { get; set; }
